Reorganise files, consolidating to a top level lib
