+++ src/core/ContractRepository.ts
@@ -1,0 +1,4 @@
+import { Contract } from '.'
+import { Repository } from '../lib'
+
+export type ContractRepository = Repository<Contract>
