+++ src/core/PortAllocationRepository.ts
@@ -1,0 +1,4 @@
+import { PortAllocation } from '.'
+import { Repository } from '../lib'
+
+export type PortAllocationRepository = Repository<PortAllocation>
