+++ src/core/index.ts
@@ -1,7 +1,8 @@
-export { Bundle, ExecutableType } from './Bundle'
-export { Contract } from './Contract'
-export { ContractExecutionInstruction } from './ContractExecutionInstruction'
-export { Engine } from './Engine'
-export { EngineClient } from './EngineClient'
-export { Events } from './Events'
-export { PortAllocation } from './PortAllocation'
+export * from './Bundle'
+export * from './Contract'
+export * from './ContractExecutionInstruction'
+export * from './ContractRepository'
+export * from './Engine'
+export * from './EngineClient'
+export * from './Events'
+export * from './PortAllocation'
