--- src/core/lib/index.ts
@@ -1,2 +1,0 @@
-export { Entity } from './Entity'
-export { NumberRange } from './NumberRange'
