+++ src/lib/Entity.ts
