+++ src/lib/NetworkInterface.ts
