+++ src/lib/PortMapper.spec.ts
@@ -1,10 +1,10 @@
 import { expect, use } from 'chai'
 import * as chaiAsPromised from 'chai-as-promised'
-import { PortAllocation } from '../../core'
-import { AllPortsAllocated } from '../../core/errors'
-import { PortMapper } from '.'
-import { InMemoryRepository } from '../infrastructure'
-import { checkPortIsFree, RogueService } from '../test'
+import { PortAllocation } from '../core'
+import { AllPortsAllocated } from '../core/errors'
+import { PortMapper } from '../lib'
+import { InMemoryRepository } from '../server/infrastructure'
+import { checkPortIsFree, RogueService } from '../server/test'
 
 use(chaiAsPromised)
