+++ src/server/application/Api.spec.ts
@@ -4,10 +4,10 @@ import * as express from 'express'
 import axios from 'axios'
 import { PubSub } from 'apollo-server'
 import { Engine, PortAllocation } from '../../core'
+import { expressEventPromiseHandler, httpEventPromiseHandler, PortMapper } from '../../lib'
 import { HttpTarGzBundleFetcher, InMemoryRepository, StubEngineClient } from '../infrastructure'
-import { Api, ContractApiServerController, ContractController, PortMapper } from '.'
-import { listen } from '../lib/express'
-import { close } from '../lib/http'
+import { Api, ContractApiServerController, ContractController } from '.'
+
 import { checkPortIsFree, populatedContractRepository } from '../test'
 const nock = require('nock')
 
@@ -43,9 +43,9 @@ describe('Api', () => {
       pubSubClient
     })
 
-    apiServer = await listen(api.app, API_PORT)
+    apiServer = await expressEventPromiseHandler.listen(api.app, API_PORT)
     const testContractApp = express()
-    testContractServer = await listen(testContractApp, 8082)
+    testContractServer = await expressEventPromiseHandler.listen(testContractApp, 8082)
     apiServerController.servers.set('abcd', testContractServer)
     nock('http://localhost:8082')
       .get('/graphql/')
@@ -53,8 +53,8 @@ describe('Api', () => {
   })
 
   afterEach(async () => {
-    await close(apiServer)
-    await close(testContractServer)
+    await httpEventPromiseHandler.close(apiServer)
+    await httpEventPromiseHandler.close(testContractServer)
     return nock.cleanAll()
   })
