+++ src/server/application/Api.ts
@@ -1,10 +1,9 @@
 import * as express from 'express'
 import { ApolloServer } from 'apollo-server-express'
 import { gql, PubSubEngine } from 'apollo-server'
-import { Bundle, Contract, Events } from '../../core'
-import { ContractApiServerController, ContractController } from '.'
-import { ContractRepository } from './lib/ContractRepository'
 import net from 'net'
+import { Bundle, Contract, ContractRepository, Events } from '../../core'
+import { ContractApiServerController, ContractController } from '.'
 import { ContractNotLoaded } from './errors'
 const httpProxy = require('http-proxy')
