+++ src/server/application/ContractApiServerController.spec.ts
@@ -4,8 +4,8 @@ import axios from 'axios'
 import { IExecutableSchemaDefinition } from 'apollo-server'
 import { PortAllocation } from '../../core'
 import { AllPortsAllocated } from '../../core/errors'
-import { ContractApiServerController, PortMapper, PortMapperConfig } from '.'
-import { Repository } from './lib/Repository'
+import { Repository, PortMapper, PortMapperConfig } from '../../lib'
+import { ContractApiServerController } from '.'
 import { InMemoryRepository, StubEngineClient } from '../infrastructure'
 import { checkPortIsFree, RogueService, testContracts } from '../test/'
