+++ src/server/application/ContractController.spec.ts
@@ -3,9 +3,9 @@ import { spy } from 'sinon'
 import * as sinonChai from 'sinon-chai'
 import axios from 'axios'
 import { PubSub } from 'apollo-server'
-import { Contract, Engine, EngineClient, PortAllocation } from '../../core'
-import { BundleFetcher, ContractApiServerController, ContractController, PortMapper } from '.'
-import { ContractRepository } from './lib/ContractRepository'
+import { Contract, ContractRepository, Engine, EngineClient, PortAllocation } from '../../core'
+import { PortMapper } from '../../lib'
+import { BundleFetcher, ContractApiServerController, ContractController } from '.'
 import { InMemoryRepository, HttpTarGzBundleFetcher, StubEngineClient } from '../infrastructure'
 import { testContracts } from '../test'
