+++ src/server/application/ContractController.ts
@@ -1,7 +1,6 @@
 import { PubSubEngine } from 'apollo-server'
-import { Contract, Engine, EngineClient } from '../../core'
+import { Contract, ContractRepository, Engine, EngineClient } from '../../core'
 import { BundleFetcher, ContractApiServerController, ContractInteractionController } from '.'
-import { ContractRepository } from './lib/ContractRepository'
 const requireFromString = require('require-from-string')
 
 type Config = {
