+++ src/server/application/Server.spec.ts
@@ -3,8 +3,8 @@ import * as chaiAsPromised from 'chai-as-promised'
 import axios from 'axios'
 import { PubSub } from 'apollo-server'
 import { Contract, Engine, PortAllocation } from '../../core'
+import { Repository } from '../../lib'
 import { Server } from '.'
-import { Repository } from './lib/Repository'
 import { InMemoryRepository, HttpTarGzBundleFetcher, StubEngineClient } from '../infrastructure'
 import { ServiceApiClient, testContracts, checkPortIsFree } from '../test'
 const nock = require('nock')
