+++ src/server/application/index.ts
@@ -3,5 +3,4 @@ export { BundleFetcher } from './BundleFetcher'
 export { ContractApiServerController } from './ContractApiServerController'
 export { ContractController } from './ContractController'
 export { ContractInteractionController } from './ContractInteractionController'
-export { PortMapper, Config as PortMapperConfig } from './PortMapper'
 export { Server } from './Server'
