--- src/server/application/lib/ContractRepository.ts
@@ -1,4 +1,0 @@
-import { Contract } from '../../../core'
-import { Repository } from './Repository'
-
-export type ContractRepository = Repository<Contract>
