--- src/server/application/lib/PortAllocationRepository.ts
@@ -1,4 +1,0 @@
-import { PortAllocation } from '../../../core'
-import { Repository } from './Repository'
-
-export type PortAllocationRepository = Repository<PortAllocation>
